# API.yaml
...
openapi: 3.0.0
info:
  title: BlogLite
  description: This is a simple API
  contact:
    email: arvind@bloglite.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: http://localhost:8080/
  description: Frontend
- url: http://127.0.0.1:5000
  description: Backend
...
paths:
...
  /api/posts:
    get:
      tags:
      - posts
      summary: get all posts
      description: getting all posts for home page
      operationId: getPosts
      responses:
        "200":
          description: Successfully Fetched All Posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
        "400":
          description: Invalid Parameters
...
  /api/activity:
    get:
      tags:
      - activity
      summary: get user activity
      description: "Get login, post, follow and unfollow activity"
      operationId: userActivity
      responses:
        "200":
          description: Successfully Made Graphs
        "400":
          description: incorrect parameters
...
components:
  schemas:
    User:
      required:
      - active
      - activity
      - email
      - followers
      - followings
      - fs_uniquifier
      - id
      - name
      - password
      - posts
      - username
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: arvind
        email:
          type: string
          example: arvind@gmail.com
        name:
          type: string
          example: Arvind Raghavendran
        phone:
          type: integer
          example: 1234567890
        bio:
          type: string
          example: Hello
        password:
          type: string
          example: "1234"
        active:
          type: boolean
          example: true
        fs_uniquifier:
          type: string
          format: utf-8
          example: b247f21163e613fe6edf222984f7c426
        posts:
          $ref: '#/components/schemas/Posts'
        followers:
          $ref: '#/components/schemas/Follow'
        followings:
          $ref: '#/components/schemas/Follow'
        activity:
          $ref: '#/components/schemas/Activity'
    Posts:
      required:
      - author
      - body
      - date
      - id
      - img_url
      - title
      - user_id
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Title of my post
        body:
          type: string
          example: Caption of my post
        date:
          type: string
          format: date-time
        user_id:
          type: integer
          example: 1
        author:
          $ref: '#/components/schemas/User'
    Follow:
      required:
      - follower_id
      - following_id
      - id
      - is_follow
      - is_following
      type: object
      properties:
        id:
          type: integer
          example: 1
        follower_id:
          type: integer
          example: 1
        following_id:
          type: integer
          example: 1
        is_follow:
          $ref: '#/components/schemas/User'
        is_following:
          $ref: '#/components/schemas/User'
    Activity:
      required:
      - active_user
      - activity
      - id
      - timestamp
      - user_id
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        activity:
          type: string
          example: login
        timestamp:
          type: string
          format: date-time
        active_user:
          $ref: '#/components/schemas/User'
...
    inline_response_200:
      type: object
      properties:
        posts:
          type: array
          items:
            allOf:
            - $ref: '#/components/schemas/Posts'
        usernames:
          type: array
          items:
            $ref: '#/components/schemas/inline_response_200_usernames'
...
    inline_response_200_usernames:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: arvind
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT